third_party/blink/web_tests/external/wpt/html/cross-origin-embedder-policy/credentialless/cors-or-credentialless/video.tentative.https-expected.txt

This is a testharness.js-based test.
PASS Main
PASS video same-origin + undefined
PASS video same-origin + anonymous
PASS video same-origin + use-credentials
FAIL video cross-origin + undefined assert_equals: coep:credentialless =>  expected (undefined) undefined but got (string) "cross_origin"
PASS video cross-origin + anonymous
PASS video cross-origin + use-credentials
Harness: the test ran to completion.
